# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ICT Trading Bot API — Capital.com integration
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /positions/{dealId}:
    delete:
      operationId: closePosition
      tags: [positions]
      summary: Manually close an open position
      parameters:
        - name: dealId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Position closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CloseResult"
        "404":
          description: Position not found
...
components:
  schemas:
...
    CloseResult:
      type: object
      required: [success, dealId, profit]
      properties:
        success:
          type: boolean
        dealId:
          type: string
        profit:
          type: number
        message:
          type: string
...